Add event_path, fix no-path bug
